Refactor using BaseTest
