+++ src/test/java/org/example/BaseTest.java
@@ -1,4 +1,51 @@
 package org.example;
 
+import com.aventstack.extentreports.ExtentReports;
+import com.aventstack.extentreports.ExtentTest;
+import com.aventstack.extentreports.Status;
+import com.aventstack.extentreports.reporter.ExtentSparkReporter;
+import org.example.pages.LoginPage;
+import org.openqa.selenium.WebDriver;
+import org.testng.ITestResult;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.BeforeTest;
+
+import java.io.IOException;
+
 public class BaseTest {
+
+    ExtentReports extent;
+    ExtentTest test;
+
+    @BeforeTest
+    public void setup(){
+        ExtentSparkReporter spark = new ExtentSparkReporter("C:/Users/2000145257/IdeaProjects/Reports/TestsReport.html");
+        extent = new ExtentReports();
+        extent.attachReporter(spark);
+    }
+
+    @AfterMethod
+    public void tearDown(ITestResult result){
+        if(result.getStatus() == ITestResult.SUCCESS){
+            test.log(Status.PASS, "Successfully passed");
+        } else if (result.getStatus() == ITestResult.FAILURE){
+            test.log(Status.FAIL, "Test failed in: " + result.getThrowable().getMessage());
+        } else if (result.getStatus() == ITestResult.SKIP){
+            test.log(Status.SKIP, "Skipped test");
+        }
+    }
+
+    @AfterTest
+    public void teardown(){
+        extent.flush();
+    }
+
+    public void navigateToApp(WebDriver driver) throws InterruptedException, IOException {
+        driver.navigate().to("https://opensource-demo.orangehrmlive.com/");
+        Thread.sleep(2000);
+        LoginPage loginPage = new LoginPage(driver);
+        loginPage.login();
+        Thread.sleep(2000);
+    }
 }
